fix-db: correccion en Dao, correccion en el tipo de dato

+++ app/src/main/java/com/example/laboratorio_pdm_c2/Dao/ArticuloDao.java
@@ -23,10 +23,10 @@ public interface ArticuloDao {
     Articulo getArticulo(int idarticulo);
 
     @Update
-    long updateArticulo(Articulo articulo);
+    int updateArticulo(Articulo articulo);
 
     @Delete
-    long deleteArticulo(Articulo articulo);
+    int deleteArticulo(Articulo articulo);
 
 
 }
